+++ cli/project/dockerfiles_test.go
@@ -151,7 +151,7 @@ RUN if id -u {{ .UserID }} 2>/dev/null; then \
         USERNAME=$(id -nu {{ .UserID }}); \
     else \
         USERNAME=cartridge; \
-        useradd -u {{ .UserID }} ${USERNAME}; \
+        useradd -l -u {{ .UserID }} ${USERNAME}; \
     fi \
     && (usermod -a -G sudo ${USERNAME} 2>/dev/null || :) \
     && (usermod -a -G wheel ${USERNAME} 2>/dev/null || :) \
@@ -188,7 +188,7 @@ RUN if id -u {{ .UserID }} 2>/dev/null; then \
         USERNAME=$(id -nu {{ .UserID }}); \
     else \
         USERNAME=cartridge; \
-        useradd -u {{ .UserID }} ${USERNAME}; \
+        useradd -l -u {{ .UserID }} ${USERNAME}; \
     fi \
     && (usermod -a -G sudo ${USERNAME} 2>/dev/null || :) \
     && (usermod -a -G wheel ${USERNAME} 2>/dev/null || :) \
@@ -235,7 +235,7 @@ RUN if id -u {{ .UserID }} 2>/dev/null; then \
         USERNAME=$(id -nu {{ .UserID }}); \
     else \
         USERNAME=cartridge; \
-        useradd -u {{ .UserID }} ${USERNAME}; \
+        useradd -l -u {{ .UserID }} ${USERNAME}; \
     fi \
     && (usermod -a -G sudo ${USERNAME} 2>/dev/null || :) \
     && (usermod -a -G wheel ${USERNAME} 2>/dev/null || :) \
@@ -272,7 +272,7 @@ RUN if id -u {{ .UserID }} 2>/dev/null; then \
         USERNAME=$(id -nu {{ .UserID }}); \
     else \
         USERNAME=cartridge; \
-        useradd -u {{ .UserID }} ${USERNAME}; \
+        useradd -l -u {{ .UserID }} ${USERNAME}; \
     fi \
     && (usermod -a -G sudo ${USERNAME} 2>/dev/null || :) \
     && (usermod -a -G wheel ${USERNAME} 2>/dev/null || :) \
@@ -309,7 +309,7 @@ func TestGetRuntimeImageDockerfileTemplateEnterprise(t *testing.T) {
 
 ### Create Tarantool user and directories
 RUN groupadd -r tarantool \
-    && useradd -M -N -g tarantool -r -d /var/lib/tarantool -s /sbin/nologin \
+    && useradd -M -N -l -g tarantool -r -d /var/lib/tarantool -s /sbin/nologin \
         -c "Tarantool Server" tarantool \
     &&  mkdir -p /var/lib/tarantool/ --mode 755 \
     && chown tarantool:tarantool /var/lib/tarantool \
@@ -355,7 +355,7 @@ RUN yum install -y zip
 
 ### Create Tarantool user and directories
 RUN groupadd -r tarantool \
-    && useradd -M -N -g tarantool -r -d /var/lib/tarantool -s /sbin/nologin \
+    && useradd -M -N -l -g tarantool -r -d /var/lib/tarantool -s /sbin/nologin \
         -c "Tarantool Server" tarantool \
     &&  mkdir -p /var/lib/tarantool/ --mode 755 \
     && chown tarantool:tarantool /var/lib/tarantool \